Avoid double scaling.
This commit factors scaling code in one place to avoid scaling twice[*].
New function `get_pixbuf_scaled` returns an icon pixbuf scaled exactly
to --icon-width, if width > 0 is specified. Otherwise the function
returns the same pixbuf that get_pixbuf returns (a YAD_ICON_BIG for
--listen or a YAD_ICON_SMALL for --compact).

[*] More precisely, when no --icon-width is given, icons are scaled zero
or one times depending on whether the icon is specified as a file path
or as a GTK icon them icon name. If --icon-width=W W > 0 is given, icons
are scaled one or two times, the second time only if the first scaling
did not produce exactly the requested width W. Typical icon sizes, such
as 16, 22, 32, 48, ... etc. are likely to only need scaling once.