9 forcefully close hung notebook tabs.
Background:

A yad notebook is implemented as a GTK multi-plug: one key process, the
--notebook owns a shared memory segment and several --plug processes,
each one a --tab, attach to the shmem segment. When the user closes the
notebook, the key send SIGUSR2 to each plug then busy waits for them
to exit, then the key releases shmem and exits itself.

This fix works around the following issue:

Repeatedly resizing a notebook with several tabs can results in some
tabs entering a process state that SIGUSR2 can't stop. For instance,
closing ([X] window button) Fatdog64 800's control panel (seven tabs)
can leave two or more background processes hung (some plugs and a key).
It is not clear why the plug processes hang in the first place.
This issue can be reproduced on GTK2 for --notebook but not for --paned.

This fix changes the way the key stops the plugs. First it sends them
SIGUSR2 as the normal case, then it waits for up to one second for all
plugs to exit. If some plug hasn't exited the key sends them SIGTERM
and waits again for another second. After which the key sends SIGTERM
again and waits for one second, and so on. The key does not exit until
all plugs have terminated.  So far no hung plug has survived a SIGTERM.

(cherry picked from commit be446b6d02d998a9969f372725eb4b8f6f3bb620)