13 **Change** --center to resize smoothly, add new --center-keep.
This commit changes the behavior of option --center, and exposes the old
behavior as new option --center-keep; --center wins over --center-keep.

Previous behavior:
- The window is initially placed in the center of the screen
- As the window is being resized it jerks very visibly
- JWM does not allow resizing it at all
- Openbox allows resizing and moves the dragged corner to the position
  of the mouse pointer when the mouse button is released
- Compiz behaves like Openbox except that if the mouse pointer is moved
  to a screen corner compiz centers the resized window.

The old behavior is generally odd-looking, and unpredictable from the
user's perspective.

New behavior:
- The window is initially placed in the center of the screen
- The window resizes smoothly
- Only the dragged corner(s) move(s) while resizing and their settle
  in place when the mouse button is released.

The new behavior looks right from a user's perspective.

The decision to tie the new behavior to an existing option (--center)
was taken in the belief that most shell script authors do not realize
the oddities of the old behavior, and assume that --center will simply
place the window in the center of the screen initially, and allow the
user to resize and move the window at will thereafter.