add color picker to color dialog and yad-tools [backport][gtk3]
Backported:
ffd6979 add color picker to color dialog and yad-tools

Notes:
* color picker only works for gtk3 build.

The intent of the upstream commit is to fix the complete lack of a
color picker in the gtk3 color dialog. Upstream leveraged the fix to
also add a screen pixel color picker to yad-tools but the implementation
is gtk3-dependent.